+++ packages/js/data/src/onboarding/selectors.ts
@@ -102,6 +102,9 @@ export type OnboardingSelectors = {
 	getTaskListsByIds: (
 		ids: string[]
 	) => ReturnType< typeof getTaskListsByIds >;
+	getTaskLists: () => ReturnType< typeof getTaskLists >;
+	getTaskList: ( id: string ) => ReturnType< typeof getTaskList >;
+	getFreeExtensions: () => ReturnType< typeof getFreeExtensions >;
 } & WPDataSelectors;
 
 export type OnboardingState = {